022 GenAI Course/Python/Week13_14/trial1/prototype_package/requirements.txt
numpy
scikit-learn
sentence-transformers
pillow
requests
